backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
llama-index==0.12.52
langchain==0.3.27
python-dotenv==1.1.1
pypdf==5.9.0
